refactor(goosecracker): move harness + agent guest-init to projects/firecracker/goosecracker
