+++ projects/firecracker/goosecracker/guest-init/cmd/loopback_other.go
