+++ projects/firecracker/goosecracker/guest-init/cmd/main_test.go
