+++ projects/firecracker/goosecracker/guest-init/internal/harness/harness.go
