+++ projects/firecracker/goosecracker/guest-init/internal/harness/harness_test.go
