+++ projects/firecracker/goosecracker/guest-init/internal/idle/idle.go
