+++ projects/firecracker/goosecracker/guest-init/internal/idle/idle_test.go
