+++ projects/firecracker/goosecracker/guest/recipes/agent.yaml
