+++ projects/firecracker/goosecracker/guest/recipes/artifact.yaml
