+++ projects/firecracker/goosecracker/guest/recipes/docs.yaml
