+++ projects/firecracker/goosecracker/guest/recipes/research.yaml
